# app/routers/rentals.py
"""
Rentals endpoints
Handles all rental/booking operations for storage units
"""
from fastapi import APIRouter, HTTPException, status, Query
from typing import List, Optional
from ..models import (
    Rental,
    RentalCreate,
    RentalUpdate,
    RentalWithDetails,
    RentalStatus
)

router = APIRouter(prefix="/api/v1/rentals", tags=["rentals"])


@router.get("/", status_code=status.HTTP_501_NOT_IMPLEMENTED)
def list_rentals(
    skip: int = Query(0, ge=0, description="Number of records to skip"),
    limit: int = Query(100, ge=1, le=100, description="Maximum number of records to return"),
    status_filter: Optional[RentalStatus] = Query(None, description="Filter by rental status")
):
    """
    Retrieve list of all rentals with optional filters
    """
    return {"detail": "Not implemented"}


@router.post("/", status_code=status.HTTP_501_NOT_IMPLEMENTED)
def create_rental(rental: RentalCreate):
    """
    Create a new rental (book a storage unit)
    """
    return {"detail": "Not implemented"}


@router.get("/{rental_id}", status_code=status.HTTP_501_NOT_IMPLEMENTED)
def get_rental(rental_id: str):
    """
    Retrieve a specific rental by ID with storage unit details
    """
    return {"detail": "Not implemented"}


@router.put("/{rental_id}", status_code=status.HTTP_501_NOT_IMPLEMENTED)
def update_rental(rental_id: str, rental: RentalUpdate):
    """
    Update an existing rental
    """
    return {"detail": "Not implemented"}


@router.delete("/{rental_id}", status_code=status.HTTP_501_NOT_IMPLEMENTED)
def delete_rental(rental_id: str):
    """
    Delete/cancel a rental
    """
    return {"detail": "Not implemented"}


@router.post("/{rental_id}/confirm", status_code=status.HTTP_501_NOT_IMPLEMENTED)
def confirm_rental(rental_id: str):
    """
    Confirm a pending rental
    """
    return {"detail": "Not implemented"}


@router.post("/{rental_id}/activate", status_code=status.HTTP_501_NOT_IMPLEMENTED)
def activate_rental(rental_id: str):
    """
    Activate a confirmed rental (when rental period starts)
    """
    return {"detail": "Not implemented"}


@router.post("/{rental_id}/complete", status_code=status.HTTP_501_NOT_IMPLEMENTED)
def complete_rental(rental_id: str):
    """
    Complete an active rental (when rental period ends)
    """
    return {"detail": "Not implemented"}


@router.post("/{rental_id}/cancel", status_code=status.HTTP_501_NOT_IMPLEMENTED)
def cancel_rental(rental_id: str):
    """
    Cancel a rental
    """
    return {"detail": "Not implemented"}


# Integration endpoints for other microservices

@router.get("/renter/{renter_user_id}", status_code=status.HTTP_501_NOT_IMPLEMENTED)
def get_rentals_by_renter(
    renter_user_id: int,
    active_only: bool = Query(False, description="Return only active rentals")
):
    """
    Get all rentals for a specific renter (renter_user_id references users.user_id)
    """
    return {"detail": "Not implemented"}


@router.get("/owner/{owner_user_id}", status_code=status.HTTP_501_NOT_IMPLEMENTED)
def get_rentals_by_owner(
    owner_user_id: int,
    active_only: bool = Query(False, description="Return only active rentals")
):
    """
    Get all rentals for storage units owned by a specific user (owner_user_id references users.user_id)
    """
    return {"detail": "Not implemented"}


@router.get("/storage-unit/{storage_unit_id}", status_code=status.HTTP_501_NOT_IMPLEMENTED)
def get_rentals_by_storage_unit(storage_unit_id: str):
    """
    Get all rentals for a specific storage unit
    """
    return {"detail": "Not implemented"}


@router.get("/location/{location_id}", status_code=status.HTTP_501_NOT_IMPLEMENTED)
def get_rentals_by_location(location_id: str):
    """
    Get all rentals for storage units at a specific location
    """
    return {"detail": "Not implemented"}